BUG: Python listed in install_requires
Python cannot be listed in the install_requires section of a setup.py file because pip cannot install python. It's also redundant because you needed python to run this setup file in the first place. This also causes errors with `pip check` because as mentioned, python is not installable using a python package manager.